+++ app/covid.ts
@@ -1,7 +1,7 @@
 import moment from 'moment';
 
 let test: string = "Bob";
-console.log("Hello TS LOU")
+console.log("Hello TS ABI")
 let abi = test;
