+++ index.html
@@ -13,8 +13,8 @@
 </head>
 <body>
 <div class="container">
-    <h6>COVID-19 Related Deaths in Scotland</h6>
+    <h6>COVID-19 Related Deaths in Scotlandxxx</h6>
 </div>
-<script type="text/javascript" src="covid.ts"></script>
+<script type="text/javascript" src="app/covid.ts"></script>
 </body>
 </html>
