+++ tests/covid.test.ts
@@ -1,4 +1,4 @@
-import { getDate } from "./covid"
+import { getDate } from "../app/covid"
 
 test('adds 1 week to start date', () => {
   expect(getDate(1)).toBe("16-Mar-2020");
